♻️ change call-service row to button row
