+++ src/panels/lovelace/create-element/create-row-element.ts
@@ -4,6 +4,7 @@ import "../entity-rows/hui-script-entity-row";
 import "../entity-rows/hui-sensor-entity-row";
 import "../entity-rows/hui-text-entity-row";
 import "../entity-rows/hui-toggle-entity-row";
+import "../special-rows/hui-button-row";
 import "../special-rows/hui-call-service-row";
 import { EntityConfig } from "../entity-rows/types";
 import { createLovelaceElement } from "./create-element-base";
@@ -15,6 +16,7 @@ const ALWAYS_LOADED_TYPES = new Set([
   "sensor-entity",
   "text-entity",
   "toggle-entity",
+  "button-row",
   "call-service",
 ]);
 const LAZY_LOAD_TYPES = {
